feat(checkout): improve shipping logic and expose default methods
