+++ libs/checkout-delivery-options/src/utils/index.ts
@@ -1,6 +1,8 @@
 export * from './debounce';
 export * from './deliveryOptionsIsRendered';
+export * from './fetchCheckoutContext';
 export * from './getDeliveryOptionsAddress';
+export * from './getPackageTypeFromShippingMethod';
 export * from './getResolvedSettings';
 export * from './shippingMethodHasDeliveryOptions';
 export * from './toggleDeliveryOptions';
